backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
gunicorn==21.2.0
python-dotenv==1.0.0
reportlab==4.0.7
